app/requirements.txt
flask==3.0.0
prometheus-flask-exporter==0.23.1
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-flask==0.44b0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
